See the documentation below!
When loading a class via a ClassLoader - you must provide the fully qualified name of the class you want to load. 
For instance, if you want to load the class Random from the package java.util - you must provide the name as java.util.Random


All system class loaders use the fully qualified name to load the corresponding classes. See the official documentation below:

http://docs.oracle.com/javase/7/docs/api/java/lang/ClassLoader.html#loadClass%28java.lang.String%29

If you want to make your custom class loader to use the path of an entry in the concrete jar you are writing a class loader for - then if you want to search if the class is loaded by a system class loader, you must give the name to the system class loader with dots instead of slashes. Otherwise an error will be raised by the system class loaders you are searching for the class in.

A good practice is to write your custom class loader in compliance with the convention for input parameters accepted in all class loaders :

use dots instead of slashes [; 

In order your class loader to be usable as all others if exported. [;